Migrate to params test
